fix test failures
